requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
databases==0.6.2
fastapi==0.95.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
pydantic==1.10.4
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
yarl==1.9.2
